Refactored VideoCapture, basic app code moved to common reprostim::CaptureApp class #61
